+++ app/Http/Controllers/CategoryController.php
@@ -8,14 +8,17 @@ use App\Http\Requests\Category\Request as CategoryRequest;
 
 class CategoryController extends Controller
 {
-    public function __construct(){
+    public function __construct()
+    {
         $this->model = Category::class;
         $this->validator = CategoryRequest::class;
     }
 
     protected function beforeDeleteActions($post)
     {
-        $post->posts()->detach();
+        if (!is_null($post->posts())) {
+            $post->posts()->detach();
+        }
         return $post;
     }
 }
